+++ packages/cli/src/types/user-config.ts
@@ -36,6 +36,6 @@ export interface UserConfig {
     noCheck: boolean
     /** Fix Inheritance and implementation type conflicts */
     fixConflicts: boolean
-    /** Wheter the project has DOM lib enabled in tsconfig/jsconfig */
-    hasDOMLib: boolean
+    /** Disables the generation of types that are in conflict with the DOM types */
+    noDOMLib: boolean
 }